pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "mgc"
version = "0.1.0"
description = "Music Generator Company - autonomous pipeline MVP"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
